Able to add assets to a scene at runtime.

+++ Src/DeviceHandlers/TextDeviceData.h
@@ -12,14 +12,14 @@ class TextDeviceData {
         void OnDeviceLost() noexcept { m_pSpriteFont.reset(); }
 
     public:
-        std::uint32_t GetHeapIndex() const noexcept { return m_descriptorHeapIndex; }
+        std::uint8_t GetHeapIndex() const noexcept { return m_descriptorHeapIndex; }
         DirectX::SpriteFont* GetSpriteFont() noexcept { return m_pSpriteFont.get(); }
 
-        void SetHeapIndex(std::uint32_t heapIndex) noexcept { m_descriptorHeapIndex = heapIndex; }
+        void SetHeapIndex(std::uint8_t heapIndex) noexcept { m_descriptorHeapIndex = heapIndex; }
         void SetSpriteFont(std::unique_ptr<DirectX::SpriteFont> pSpriteFont) noexcept { m_pSpriteFont = std::move(pSpriteFont); }
 
     private:
-        std::uint32_t m_descriptorHeapIndex;
+        std::uint8_t m_descriptorHeapIndex;
         std::unique_ptr<DirectX::SpriteFont> m_pSpriteFont;
 };
