+++ Src/DeviceHandlers/TextureDeviceData.h
@@ -11,14 +11,14 @@ class TextureDeviceData {
         void OnDeviceLost() noexcept { m_pTexture.Reset(); }
 
     public:
-        std::uint32_t GetHeapIndex() const noexcept { return m_desciptorHeapIndex; }
+        std::uint8_t GetHeapIndex() const noexcept { return m_desciptorHeapIndex; }
 
         Microsoft::WRL::ComPtr<ID3D12Resource>& GetTexture() noexcept { return m_pTexture; }
 
-        void SetHeapIndex(std::uint32_t heapIndex) noexcept { m_desciptorHeapIndex = heapIndex; }
+        void SetHeapIndex(std::uint8_t heapIndex) noexcept { m_desciptorHeapIndex = heapIndex; }
 
     private:
-        std::uint32_t m_desciptorHeapIndex;
+        std::uint8_t m_desciptorHeapIndex;
         Microsoft::WRL::ComPtr<ID3D12Resource> m_pTexture;        
 };
